refactor(qa): keep compaction events with response builders
